% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverage}
\alias{coverage}
\title{Simulated dataset for Vaccination coverage for DTap vaccine}
\format{
A dataframe with 6 rows and 2 columns:
\describe{
\item{x}{DTap-3 vaccination coverage, in percentage}
\item{y}{DTap-4 vaccination coverage, in percentage}
}'
}
\usage{
coverage
}
\description{
A dataset containing the 6 possible immunization coverage of 3rd and 4th
dose of DTap vaccines in US
}
\keyword{datasets}
